Launch OpenStack with Chef on Packet.net [#103][#102]
Terraform + Ansible -> Packet.net works.  And works for multiple nodes (4 at
the moment).  The first note is defined as the system controller, and
has control logic to launch ansible configuration to enable the deployment.

Additional roles were created to deploy chef, chef-client, ntp, and the
original ansible was updated to support RedHat (as that was a closer model
to most Chef based deployments where we had some experience).

The README.md is an initial document that needs more work to cover the
process in more detail.