Resolve hugepage interaction issues with vpp and KVM
Fixes Issue #101

Seems that the principal interaction failure was due to a major
mismatch between VPP's hugepage config (or lack thereof), and
a hugepage config that also works with OpenStack.  Once enough
pages were available to VPP (more than 512) and once the now
default OS based hugepage directory (/dev/hugepages) was set
it was possible to launch VMs and have Neutron properly configure
the environment.

Additional errors that caused issues:
VPP and Neutron DHCP do not play nicely, or at least not with this
specific cnfiguration, and the network script that helps set up a
test network needed the --no-dhcp option in addition to the current
--gateway none confguration.

The neutron flavor script needed to include a hw:numa_nodes=1 or
greater parameter or the cpu and memory specific parameters are
ignored. It may be that these are no longer necessary with the
hugepage config set on VPP, but this does effectively pin
the VM to one NUMA node or the other. It's not clear that the
config does a 1 to 1 mapping given the OpenStack documentation
but it does appear to configure a libvirt cpu pinning operation
with this specific set of flavor options.

VPP branch from git 'stable/18.07' still has active commits, so
the version was locked to a specific git commit id to ensure
repeatable results. The above mentioned DHCP issue was the
apparent cause of inconsistences rather than the updated
commit to VPP, but at this point, we're commit locked.

VPP-agent (the neutron-agent for host config) was not starting
due to a reliance on the neutron-server service being started
but the neutron-servrer service doesn't exist on the compute
nodes, where the agent normally needs to run (router/dhcp configs
may still require the agent on the OpenStack control nodes
as well).

VPP Startup.conf was simplified but the hugepage parameter
(which doesn't exist in the documentation) was added
to point to the proper hugepages directory.

Grub for openstack was also updated to ensure that there
are enough hugepages for a reasonable number of VMs per host
at least for testing purposes.