Move to current 'master' OpenStack-Chef codebase
Addresses Issues [#189] and [#192]

Previous openstack-chef code was locked to stable/queens, but
that version doesn't specify the version of the apache2 cookbook.
Apparently the just released apache2 v6.0.0 moved the entire
cookbook to using a new "wrapper" style of cookbook model but
the OpenStack code doesn't yet support this.  The master
branch code and more specifically commit:
 3faee3507dcb54c8986d572e5d12fee68a39dfa7
does.

In addition, we added etcd to the latest v3 (3.3.12)