WIP Roadmap
This document defines a high level roadmap for the CNF Testbed initiative.

The roadmap below is only a selection of some of the major features we plan to implement in the near future. To get a more complete overview of planned features and current work see the [issue tracker](https://github.com/cncf/cnf-testbed/issues) and [milestones](https://github.com/cncf/cnf-testbed/milestones) in GitHub.

The dates below should not be considered authoritative, but rather indicative of the projected timeline of the project.